put modal in its own file
