+++ components/pages/submission-system/program-clinical-submission/PageContent.tsx
@@ -43,7 +43,8 @@ import { displayDateAndTime } from 'global/utils/common';
 import { capitalize } from 'global/utils/stringUtils';
 import { useToaster } from 'global/hooks/toaster';
 import ErrorNotification, { getDefaultColumns } from '../ErrorNotification';
-import { ModalPortal, useGlobalLoader } from 'components/ApplicationRoot';
+import { useGlobalLoader } from 'components/ApplicationRoot';
+import ModalPortal from 'components/Modal';
 import SignOffValidationModal from './SignOffValidationModal';
 import SubmissionSummaryTable from './SubmissionSummaryTable';
 import useUserConfirmationModalState from './useUserConfirmationModalState';
